generalized the link fixer cog and added more domains
